Refactor admin dashboard user table
