+++ src/components/dashboard/admin/AdminDashboard.tsx
@@ -7,52 +7,28 @@ import {
   Card,
   Grid,
   CircularProgress,
-  TextField,
-  InputAdornment,
   Select,
-  Button,
   MenuItem,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Paper,
-  IconButton,
-  Chip,
-  Pagination,
-  FormControl,
   SelectChangeEvent,
   Tooltip,
-  TableFooter,
 } from "@mui/material";
+
 import {
-  DataGridPremium,
-  GridColDef,
-  GridPaginationModel,
-} from "@mui/x-data-grid-premium";
-import {
-  Search as SearchIcon,
   Info as InfoIcon,
-  MoreVert as MoreVertIcon,
-  FileDownloadOutlined as FileDownloadOutlinedIcon,
   ChevronLeft as ChevronLeftIcon,
   ChevronRight as ChevronRightIcon,
   ExpandMore,
   InfoOutlined,
 } from "@mui/icons-material";
 import DashboardContent from "../DashboardContent";
+import AdminUsersTable from "./AdminUsersTable";
 import { useAuth } from "../../../context/AuthContext";
-import { fetchDivisions, fetchDepartments } from "../../../services/suggestions";
+
 import {
-  AdminDashboardUserActivityResponse,
   AdminDashboardUserStatsResponse,
   UserStat,
   RoleCount,
   fetchAdminDashboardStats,
-  fetchAdminUsersTable,
-  AdminUsersTableRequest,
 } from "../../../services/admin";
 import { fetchRoles } from "../../../services/roles";
 import { DateRange } from "@mui/x-date-pickers-pro";
@@ -517,29 +493,8 @@ const AdminDashboard = () => {
   const [dashboardData, setDashboardData] = useState(adminDashboardData);
   const [dashboardStats, setDashboardStats] =
     useState<AdminDashboardUserStatsResponse | null>(null);
-  const [userActivity, setUserActivity] = useState<
-    AdminDashboardUserActivityResponse[]
-  >([]);
-
-  const [userActivityParams, setUserActivityParams] =
-    useState<AdminUsersTableRequest>({});
-
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [department, setDepartment] = useState("All Department");
-  const [division, setDivision] = useState("All Divisions");
-  const [departments, setDepartments] = useState<string[]>([]);
-  const [divisions, setDivisions] = useState<string[]>([]);
-  const [rolesList, setRolesList] = useState<string[]>([]);
-  const [isLoadingDepartments, setIsLoadingDepartments] = useState(false);
-  const [isLoadingDivisions, setIsLoadingDivisions] = useState(false);
-  const [role, setRole] = useState("All Roles");
-  const [status, setStatus] = useState("All Status");
-  const [timeframe, setTimeframe] = useState("All Time");
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(50);
-  const [dateRange, setDateRange] = useState<DateRange<Dayjs>>([null, null]);
 
   const [engagementDateRange, setEngagementDateRange] = useState<DateRange<Dayjs>>([
     dayjs().subtract(6, "day"),
@@ -551,54 +506,12 @@ const AdminDashboard = () => {
     ActiveUserMetricsHistoryItem[]
   >([]);
 
-  const [userActivityData, setUserActivityData] = useState({
-    users: adminDashboardData.userActivity,
-    total: adminDashboardData.totalUsers,
-  });
-
   const engagementDataFiltered = userEngagementData;
 
   const handleEngagementDateRangeApplyCallback = (range: DateRange<Dayjs>) => {
     setEngagementDateRange(range);
   };
 
-  const handleUserDateRangeApplyCallback = (range: DateRange<Dayjs>) => {
-    setDateRange(range);
-    setUserActivityParams({
-      ...userActivityParams,
-      start_time: range[0] ? range[0].toISOString() : undefined,
-      end_time: range[1] ? range[1].toISOString() : undefined,
-    });
-    setPage(0);
-  };
-
-  const loadUserActivity = async () => {
-    try {
-      setIsLoading(true);
-      setError(null);
-
-      const payload: AdminUsersTableRequest = {
-        page,
-        limit: rowsPerPage,
-        search: searchQuery || undefined,
-        division: division !== "All Divisions" ? division : undefined,
-        department: department !== "All Department" ? department : undefined,
-        role: role !== "All Roles" ? role : undefined,
-        status: status !== "All Status" ? status.toUpperCase() : undefined,
-        start_time: userActivityParams.start_time,
-        end_time: userActivityParams.end_time     
-      };
-
-      const data = await fetchAdminUsersTable(payload);
-      setUserActivity(data.items);
-      setUserActivityData({ users: data.items, total: data.total_count });
-    } catch (error) {
-      console.error("Error loading dashboard data:", error);
-      setError("Failed to load dashboard data");
-    } finally {
-      setIsLoading(false);
-    }
-  };
 
   const loadAdminDashboardStats = async () => {
     if (currentWorkspaceId) {
@@ -648,52 +561,9 @@ const AdminDashboard = () => {
   };
 
   useEffect(() => {
-    loadUserActivity();
     loadAdminDashboardStats();
     loadUserEngagementMetrics();
   }, [user?.id, currentWorkspaceId]);
-
-  // Load divisions and departments when workspace changes
-  useEffect(() => {
-    if (!currentWorkspaceId) return;
-
-    const loadDivisions = async () => {
-      setIsLoadingDivisions(true);
-      try {
-        const data = await fetchDivisions(currentWorkspaceId);
-        setDivisions(data);
-      } catch (err) {
-        console.error("Failed to load divisions:", err);
-      } finally {
-        setIsLoadingDivisions(false);
-      }
-    };
-
-    const loadDepartments = async () => {
-      setIsLoadingDepartments(true);
-      try {
-        const data = await fetchDepartments(currentWorkspaceId);
-        setDepartments(data);
-      } catch (err) {
-        console.error("Failed to load departments:", err);
-      } finally {
-        setIsLoadingDepartments(false);
-      }
-    };
-
-    const loadRoles = async () => {
-      try {
-        const data = await fetchRoles();
-        setRolesList(data.map((r) => r.name));
-      } catch (err) {
-        console.error("Failed to load roles:", err);
-      }
-    };
-
-    loadDivisions();
-    loadDepartments();
-    loadRoles();
-  }, [currentWorkspaceId]);
 
   useEffect(() => {
     loadUserEngagementMetrics();
@@ -732,252 +602,13 @@ const AdminDashboard = () => {
   //   rowsPerPage,
   // ]);
 
-  useEffect(() => {
-    loadUserActivity();
-  }, [userActivityParams, searchQuery, page, rowsPerPage]);
 
-  const handleChangePage = (
-    event: React.ChangeEvent<unknown>,
-    newPage: number
-  ) => {
-    setPage(newPage - 1);
-  };
-
-  const handleChangeRowsPerPage = (event: SelectChangeEvent<string>) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
-
-  const handleDepartmentChange = (event: SelectChangeEvent<string>) => {
-    setDepartment(event.target.value);
-    setUserActivityParams({
-      ...userActivityParams,
-      department: event.target.value,
-    });
-    setPage(0);
-  };
-
-  const handleDivisionChange = (event: SelectChangeEvent<string>) => {
-    setDivision(event.target.value);
-    setUserActivityParams({
-      ...userActivityParams,
-      division: event.target.value,
-    });
-    setPage(0);
-  };
-
-  const handleRoleChange = (event: SelectChangeEvent<string>) => {
-    setRole(event.target.value);
-    setUserActivityParams({
-      ...userActivityParams,
-      role: event.target.value,
-    });
-    setPage(0);
-  };
-
-  const handleStatusChange = (event: SelectChangeEvent<string>) => {
-    setStatus(event.target.value);
-    setUserActivityParams({
-      ...userActivityParams,
-      status: event.target.value,
-    });
-    setPage(0);
-  };
-
-  const handleTimeframeChange = (event: SelectChangeEvent<string>) => {
-    setTimeframe(event.target.value);
-  };
 
 
   const handleEngagementRoleChange = (event: SelectChangeEvent<string>) => {
     setEngagementRole(event.target.value);
   };
 
-  const handleExportCsv = () => {
-    const headers = [
-      "Name",
-      "Email",
-      "Role",
-      "Division",
-      "Department",
-      "Added On",
-      "Status",
-      "Assigned Simulations",
-      "Completion Rate",
-      "Adherence Rate",
-      "Average Score",
-      "Activated On",
-      "Deactivated On",
-      "Login Count",
-      "Last Login Date",
-      "Last Session Duration",
-    ];
-
-    const csvRows = [headers.join(",")];
-    userActivity.forEach((user) => {
-      const row = [
-        user.name,
-        user.email,
-        user.role,
-        user.division,
-        user.department,
-        user.addedOn,
-        user.status,
-        user.assignedSimulations,
-        user.completionRate,
-        user.adherenceRate,
-        user.averageScore,
-        user.activatedOn,
-        user.deActivatedOn,
-        user.loginCount,
-        user.lastLoginOn,
-        user.lastSessionDuration,
-      ].join(",");
-      csvRows.push(row);
-    });
-    const csvContent = csvRows.join("\n");
-    const blob = new Blob([csvContent], { type: "text/csv;charset=utf-8;" });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = "user_activity.csv";
-    link.click();
-    URL.revokeObjectURL(url);
-  };
-
-  const userColumns: GridColDef[] = [
-    {
-      field: "nameEmail",
-      headerName: "Name & Email",
-      flex: 1,
-      minWidth: 200,
-      renderCell: (params) => (
-        <Stack>
-          <Typography variant="body2" color="#000000CC" fontWeight="medium">
-            {params.row.name}
-          </Typography>
-          <Typography variant="caption" color="#00000099">
-            {params.row.email}
-          </Typography>
-        </Stack>
-      ),
-      sortable: false,
-    },
-    {
-      field: "role",
-      headerName: "Role",
-      flex: 0.5,
-      minWidth: 120,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{ bgcolor: "#F2F4F7", color: "#344054", fontWeight: "medium" }}
-        />
-      ),
-    },
-    {
-      field: "divisionDepartment",
-      headerName: "Division & Department",
-      flex: 1,
-      minWidth: 200,
-      renderCell: (params) => (
-        <Stack>
-          <Typography variant="body2">{params.row.division}</Typography>
-          <Typography variant="caption" color="text.secondary">
-            {params.row.department}
-          </Typography>
-        </Stack>
-      ),
-      sortable: false,
-    },
-    { field: "addedOn", headerName: "Added On", flex: 0.6, minWidth: 120 },
-    {
-      field: "status",
-      headerName: "Status",
-      flex: 0.5,
-      minWidth: 120,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{
-            bgcolor:
-              params.value === "ACTIVE" ? "#F2F4F7" : "#FEF3F2",
-            color: params.value === "ACTIVE" ? "#344054" : "#B42318",
-            fontWeight: "medium",
-          }}
-        />
-      ),
-    },
-    {
-      field: "assignedSimulations",
-      headerName: "Assigned Simulations",
-      flex: 0.7,
-      minWidth: 150,
-    },
-    {
-      field: "completionRate",
-      headerName: "Completion Rate",
-      flex: 0.6,
-      minWidth: 150,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{ bgcolor: "#F2F4F7", color: "#344054", fontWeight: "medium" }}
-        />
-      ),
-    },
-    {
-      field: "adherenceRate",
-      headerName: "Adherence Rate",
-      flex: 0.6,
-      minWidth: 150,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{ bgcolor: "#F2F4F7", color: "#344054", fontWeight: "medium" }}
-        />
-      ),
-    },
-    {
-      field: "averageScore",
-      headerName: "Average Score",
-      flex: 0.6,
-      minWidth: 150,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{ bgcolor: "#F2F4F7", color: "#344054", fontWeight: "medium" }}
-        />
-      ),
-    },
-    { field: "activatedOn", headerName: "Activated On", flex: 0.6, minWidth: 150 },
-    { field: "deActivatedOn", headerName: "Deactivated On", flex: 0.7, minWidth: 150 },
-    {
-      field: "loginCount",
-      headerName: "Login Count",
-      flex: 0.5,
-      minWidth: 120,
-      renderCell: (params) => (
-        <Chip
-          label={params.value}
-          size="small"
-          sx={{ bgcolor: "#F2F4F7", color: "#344054", fontWeight: "medium" }}
-        />
-      ),
-    },
-    { field: "lastLoginOn", headerName: "Last Login Date", flex: 0.7, minWidth: 150 },
-    {
-      field: "lastSessionDuration",
-      headerName: "Last Session Duration",
-      flex: 0.7,
-      minWidth: 170,
-    },
-  ];
 
 
   if (isLoading) {
@@ -1129,242 +760,7 @@ const AdminDashboard = () => {
             User Status and Activity Log
           </Typography>
 
-          <Stack>
-            {/* Search and Filters */}
-            <Stack
-              direction={{
-                sm: "column",
-                md: "row",
-              }}
-              gap={2}
-              bgcolor="#F9FAFB"
-              borderRadius={1.5}
-              p={1.5}
-              justifyContent="space-between"
-            >
-              <TextField
-                placeholder="Search by Assignment Name or ID"
-                value={searchQuery}
-                onChange={(e) => {
-                  setSearchQuery(e.target.value);
-                  setPage(0);
-                }}
-                InputProps={{
-                  startAdornment: (
-                    <InputAdornment position="start">
-                      <SearchIcon />
-                    </InputAdornment>
-                  ),
-                }}
-                sx={{
-                  bgcolor: "white",
-                  boxShadow: "0px 1px 2px 0px #1018280D",
-                  "& .MuiOutlinedInput-root": {
-                    "& fieldset": {
-                      border: "1px solid #00000014",
-                      borderRadius: 1,
-                      borderColor: "#E0E0E0",
-                    },
-                    "&:hover fieldset": {
-                      borderColor: "#C0C0C0",
-                    },
-                    "&.Mui-focused fieldset": {
-                      borderColor: "#E0E0E0",
-                    },
-                  },
-                }}
-                size="small"
-              />
-
-              <Stack
-                direction={{
-                  sm: "column",
-                  md: "row",
-                }}
-                gap={2}
-              >
-                <Select
-                  value={department}
-                  onChange={handleDepartmentChange}
-                  displayEmpty
-                  IconComponent={ExpandMore}
-                  MenuProps={menuSelectProps}
-                  sx={menuSelectsx}
-                  size="small"
-                >
-                  <MenuItem sx={menuItemSx} value="All Department">
-                    All Department
-                  </MenuItem>
-                  {departments.map((dept) => (
-                    <MenuItem sx={menuItemSx} key={dept} value={dept}>
-                      {dept}
-                    </MenuItem>
-                  ))}
-                </Select>
-
-                <Select
-                  value={division}
-                  onChange={handleDivisionChange}
-                  displayEmpty
-                  IconComponent={ExpandMore}
-                  MenuProps={menuSelectProps}
-                  sx={menuSelectsx}
-                  size="small"
-                >
-                  <MenuItem sx={menuItemSx} value="All Divisions">
-                    All Divisions
-                  </MenuItem>
-                  {divisions.map((div) => (
-                    <MenuItem sx={menuItemSx} key={div} value={div}>
-                      {div}
-                    </MenuItem>
-                  ))}
-                </Select>
-
-                <Select
-                  value={role}
-                  onChange={handleRoleChange}
-                  displayEmpty
-                  IconComponent={ExpandMore}
-                  MenuProps={menuSelectProps}
-                  sx={menuSelectsx}
-                  size="small"
-                >
-                  <MenuItem sx={menuItemSx} value="All Roles">
-                    All Roles
-                  </MenuItem>
-                  {rolesList.map((r) => (
-                    <MenuItem sx={menuItemSx} key={r} value={r}>
-                      {r}
-                    </MenuItem>
-                  ))}
-                </Select>
-
-                <Select
-                  value={status}
-                  onChange={handleStatusChange}
-                  displayEmpty
-                  IconComponent={ExpandMore}
-                  MenuProps={menuSelectProps}
-                  sx={menuSelectsx}
-                  size="small"
-                >
-                  <MenuItem sx={menuItemSx} value="All Status">
-                    All Status
-                  </MenuItem>
-                  <MenuItem sx={menuItemSx} value="Active">
-                    Active
-                  </MenuItem>
-                  <MenuItem sx={menuItemSx} value="Inactive">
-                    Inactive
-                  </MenuItem>
-                </Select>
-
-                <DateSelector
-                  dateRange={dateRange}
-                  setDateRange={setDateRange}
-                  handleDateRangeApplyCallback={handleUserDateRangeApplyCallback}
-                  variant="managerGlobal"
-                />
-                <Button
-                  variant="outlined"
-                  startIcon={<FileDownloadOutlinedIcon />}
-                  onClick={handleExportCsv}
-                >
-                  Export CSV
-                </Button>
-              </Stack>
-            </Stack>
-
-            {/* User Activity Table */}
-            <Paper
-              sx={{
-                boxShadow: "none",
-                border: "1px solid #E5E7EB",
-                borderRadius: "12px",
-                overflow: "hidden",
-                width: "100%",
-              }}
-            >
-              <DataGridPremium
-                autoHeight
-                rowHeight={68}
-                rows={userActivity}
-                columns={userColumns}
-                getRowId={(row) => row.id}
-                loading={isLoading}
-                pagination
-                paginationMode="server"
-                rowCount={userActivityData.total}
-                pageSizeOptions={[10, 25, 50, 100]}
-                paginationModel={{ page, pageSize: rowsPerPage }}
-                onPaginationModelChange={(model: GridPaginationModel) => {
-                  setPage(model.page);
-                  setRowsPerPage(model.pageSize);
-                }}
-                scrollbarSize={8}
-                sx={{
-                  border: "none",
-                  width: "100%",
-                  "& .MuiDataGrid-main": {
-                    border: "none",
-                  },
-                  "& .table-header": {
-                    backgroundColor: "#ffffff",
-                    borderBottom: "1px solid #E5E7EB",
-                    fontSize: "14px",
-                    fontWeight: 500,
-                    color: "#6B7280",
-                    "& .MuiDataGrid-columnHeaderTitle": {
-                      fontSize: "14px",
-                      fontWeight: 500,
-                      color: "#6B7280",
-                    },
-                  },
-                  "& .MuiDataGrid-columnHeaders": {
-                    borderBottom: "1px solid #E5E7EB",
-                    minHeight: "56px !important",
-                    maxHeight: "56px !important",
-                  },
-                  "& .MuiDataGrid-columnHeader": {
-                    padding: "0 16px",
-                    "&:focus, &:focus-within": {
-                      outline: "none",
-                    },
-                  },
-                  "& .MuiDataGrid-cell": {
-                    borderBottom: "1px solid #F3F4F6",
-                    padding: "12px 16px",
-                    fontSize: "14px",
-                    color: "#374151",
-                    minHeight: "68px !important",
-                    maxHeight: "68px !important",
-                    display: "flex",
-                    alignItems: "center",
-                    "&:focus, &:focus-within": {
-                      outline: "none",
-                    },
-                  },
-                  "& .MuiDataGrid-row": {
-                    backgroundColor: "#ffffff",
-                    minHeight: "68px !important",
-                    maxHeight: "68px !important",
-                    "&:nth-of-type(even)": {
-                      backgroundColor: "#F9FAFB",
-                    },
-                  },
-                  "& .MuiDataGrid-footerContainer": {
-                    borderTop: "1px solid #E5E7EB",
-                    backgroundColor: "#ffffff",
-                    minHeight: "60px",
-                  },
-                  "& .MuiDataGrid-selectedRowCount": {
-                    display: "none",
-                  },
-                }}
-              />
-            </Paper>
-          </Stack>
+          <AdminUsersTable />
         </Stack>
       </Container>
     </DashboardContent>
